Added support for train bearing from the API
* Include bearing (degrees clockwise from north) as reported by the MBTA
  API
* Apply corresponding rotation to map annotations so markers point in
  the actual direction reported for the trains

Bug in this version: by rotating the entire annotation view (by affine
transform), I inadvertently also rotated the callout bubbles.  First
attempt at a fix was to rotate the UIImage instead of the annotation
view; that worked but it broke the animation, as the change of image
from one rotation to another could not be animated.

Next fix attempt will be to put the image in a subview which can be
rotated independently of the annotation.