Moved marker into subview for independent rotation
Fixes bug in initial version of train-bearing display support where the
callout bubble was rotated along with the view.  Now the image is
contained in a UIImageView subview of the main annotation view, which
is rotated relative to the annotation view, which remains itself
unrotated.

This update also adjusts the center offset for the annotations to move
them orthogonally to the direction of travel, in effect placing them
on the right side of the track.

Bug: marker rotation still doesn't take map rotation into effect.